% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_corrected_mse.R
\name{compute_corrected_mse}
\alias{compute_corrected_mse}
\title{Compute second-order correct MSE}
\usage{
compute_corrected_mse(C_cluster, clusterID, K_inv, sig_u, sig_e)
}
\arguments{
\item{C_cluster}{Cluster-level covariates for fixed and random parameters}

\item{clusterID}{Cluster labels}

\item{K_inv}{Inverse of K matrix}

\item{sig_u}{Variance parameter of random effects}

\item{sig_e}{Variance parameter of errors}
}
\value{
List with parameters
\itemize{
\item \code{mse} First order correct MSE of mixed effects
\item \code{mse_corrected} First order correct MSE of mixed effects
}
}
\description{
Function \code{compute_mse} provides first and second order MSE estimates for
mixed parameter
}
