% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mse.R
\name{compute_mse}
\alias{compute_mse}
\title{Compute MSE}
\usage{
compute_mse(C_cluster, K_inv)
}
\arguments{
\item{C_cluster}{Cluster-level covariates for fixed and random parameters}

\item{K_inv}{Inverse of K matrix}
}
\value{
mse MSE of mixed effects
}
\description{
Function \code{compute_mse} provides first order correct MSE for mixed effects
}
