% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Sigma.R
\name{compute_Sigma}
\alias{compute_Sigma}
\title{Compute Sigma}
\usage{
compute_Sigma(X, invR, invV)
}
\arguments{
\item{X}{Matrix with covariates for fixed effects}

\item{invR}{Inverse of covariance matrix R of errors}

\item{invV}{Inverse of covariance matrix V of response variable}
}
\value{
List with elements:
\itemize{
\item \code{Sigma} - matrix Sigma
\item \code{sqrt_invxVx} - square root of variance matrix of fixed effects
}
}
\description{
Function \code{compute_Sigma} computes matrix Sigma
}
\keyword{internal}
