% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_upsilon.R
\name{compute_upsilon}
\alias{compute_upsilon}
\title{Compute matrix upsilon}
\usage{
compute_upsilon(modelset, p, common = NULL)
}
\arguments{
\item{modelset}{Type of model set:
\itemize{
\item \code{"all_subsets"} - all possible models for all parameters
\item \code{"nested"} - a sequence of nested models for all parameters
\item \code{"part_subset"} - all possible models for a subset of parameters
}}

\item{p}{Number of all fixed parameters under consideration (intercept included)}

\item{common}{A vector indicating variables forced to be present in each model}
}
\description{
Function \code{compute_upsilon} computes extended selection matrix upsilon
}
\details{
\itemize{
\item \code{modelset = "nested"}
The first model contains one covariate
from the first column of X, then two covariates from the first and the second column
of X, etc.
*\code{common}
A vector of length at most p, consisting of those column numbers of X for which the
variables are forced to be present in every model (e.g. c(2,5) for variables 2 and 5).
If provided, it is used in model.set="partsubsets".
}
}
\examples{

upsilon_all_subset = compute_upsilon(modelset = "all_subsets", p = 5)

}
