% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_modelset.R
\name{create_modelset}
\alias{create_modelset}
\title{Create matrix with all models}
\usage{
create_modelset(
  modelset = c("all_subsets", "nested", "part_subset"),
  common = NULL,
  p = 5
)
}
\arguments{
\item{modelset}{Type of model set:
\itemize{
\item \code{"all_subsets"} - all possible models for all parameters
\item \code{"nested"} - a sequence of nested models for all parameters
\item \code{"part_subset"} - all possible models for a subset of parameters
}}

\item{common}{A vector indicating variables forced to be present in each model}

\item{p}{Number of all fixed parameters under consideration (intercept included)}
}
\value{
modelset_matrix Matrix composed of zeros and ones. Ones correspond to
parameters in a model which is represented in nth row.
}
\description{
Function \code{create_modelset} creates a matrix indicating
covariates (fixed effects) in each model considered by the user
}
\details{
\itemize{
\item \code{modelset = "nested"}
The first model contains one covariate
from the first column of X, then two covariates from the first and the second column
of X, etc.
*\code{common}
A vector of length at most p, consisting of those column numbers of X for which the
variables are forced to be present in every model (e.g. c(2,5) for variables 2 and 5).
If provided, it is used in model.set="partsubsets".
}
}
\examples{
#"all_subsets"
modelset  = create_modelset("all_subsets", p = 5)

#"nested"
modelset  = create_modelset("nested", p = 5)

#"part_subset"
modelset  = create_modelset("part_subset", p = 5,
                             common = c(1:3))

}
