% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Z.R
\name{create_Z}
\alias{create_Z}
\title{Create matrix Z}
\usage{
create_Z(
  model = c("NERM", "FHM", "RIRS"),
  clusterID,
  nrandom = 1,
  X = NULL,
  intercept = TRUE
)
}
\arguments{
\item{model}{Type of mixed model: NERM, FHM, RIRS (random slopes and random intercepts)}

\item{clusterID}{Vector with cluster labels}

\item{nrandom}{Which slopes should be random? Default: \code{1}}

\item{X}{Matrix with covariates for fixed effects. Default: \code{NULL}}

\item{intercept}{Is column of ones, representing the intercept, present in X
Default: \code{TRUE}}
}
\value{
\item{Z}{Matrix of covariates for random effects}
}
\description{
Function \code{create_Z} creates matrix Z which contains
covariates of random effects
}
\details{
Matrix X is only needed if we select an option \code{model = "RIRS"}.
}
\examples{

#NERM
clusterID = rep(1:5, 2:6)
Z = create_Z("NERM", clusterID)

#FHM
clusterID = c(1:5)
Z = create_Z("FHM", clusterID)

#RIRS
clusterID = rep(1:5, 10)
n = 50
p = 10
X = matrix(rnorm(n * p), n, p)
Z = create_Z("RIRS", clusterID, nrandom = 4,
             X, intercept = FALSE)

}
