% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_NERM.R
\name{estimate_NERM}
\alias{estimate_NERM}
\title{Estimate nested error regression model (NERM)}
\usage{
estimate_NERM(X, y, clusterID, X_cluster = NULL)
}
\arguments{
\item{X}{Matrix with covariates for fixed effects}

\item{y}{Vector of responses}

\item{clusterID}{Vector with cluster labels}

\item{X_cluster}{Matrix with cluster level covariates for fixed effects.
Default: \code{NULL}.}
}
\value{
List with parameters:
\item{fit_model_fixed}{Estimated model using fixed effects}
\item{fit_model_mixed}{Estimated model using fixed and random effects}
\item{sig_u}{Variance parameter of random effects}
\item{sig_e}{Variance parameter of errors}
\item{beta}{Fixed effects (regression parameters)}
\item{u}{Random effects}
\item{mu}{Mixed effects}
}
\description{
Estimate NERM, store fixed and random parameters and variances
}
\examples{
n = 10
m_i = 5
m_total = 50

clusterID = rep(1:5, n)
p = 10
beta = rep(2, p)
u_i = rnorm(n, 0, 2)
u_i_aug = rep(u_i, each = m_i)
X = matrix(rnorm(m_total * p), m_total, p)
y = X\%*\%beta + u_i_aug + rnorm(m_total, 0, 1)
fit_nerm <- estimate_NERM(X = X, y = y,
                          clusterID = clusterID,
                          X_cluster = NULL)

}
