% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_starting_points.R
\name{find_starting_points}
\alias{find_starting_points}
\title{Find starting points}
\usage{
find_starting_points(
  n_starting_points,
  p,
  n_models_to_compare,
  list_constraints,
  scale_mvrnorm = 10
)
}
\arguments{
\item{n_starting_points}{Number of initial starting points for sampling from truncated distribution}

\item{p}{Number of all fixed parameters under consideration (intercept included)}

\item{n_models_to_compare}{Number of models in the model set to compare with}

\item{list_constraints}{List wich describes the constraints imposed on the normal distribution}

\item{scale_mvrnorm}{Scale parameter for multivariate normal distribution to sample}
}
\description{
Function \code{find_starting_points} finds starting points
to sample from constrained normal distribution
}
