% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results.R
\name{format_results}
\alias{format_results}
\title{Format results to plot}
\usage{
format_results(x, y, z, type = c("regular", "corrected", "both"))
}
\arguments{
\item{x}{An object of class \code{naive_CI}}

\item{y}{An object of class \code{postcAIC_CI}}

\item{z}{An object of class \code{postOBSP_CI}}

\item{type}{Type of naive CI (using first order, second order or
both MSE estimators)}
}
\description{
Format results to plot
}
