% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multi_CI.R
\name{plot.multi_CI}
\alias{plot.multi_CI}
\title{Plots naive, post-cAIC and post-OBSP CI for mixed effects}
\usage{
\method{plot}{multi_CI}(
  x,
  col = NULL,
  xlab = NULL,
  ylab = NULL,
  type = c("regular", "corrected", "both"),
  order_estimates = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{multi_CI}}

\item{col}{Colour of CI}

\item{xlab}{Label for the x axis}

\item{ylab}{Label for the y axis}

\item{type}{Type of CI (using first order, second order or
both MSE estimators)}

\item{order_estimates}{Order of intervals in the plot}

\item{...}{Additional parameters}
}
\description{
Plots naive, post-cAIC and post-OBSP CI for mixed effects
}
