% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.naive_CI.R
\name{plot.naive_CI}
\alias{plot.naive_CI}
\title{Plots naive confidence intervals (CI) for mixed effects}
\usage{
\method{plot}{naive_CI}(
  x,
  col = NULL,
  xlab = NULL,
  ylab = NULL,
  type = c("regular", "corrected", "both"),
  order_estimates = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{naive_CI}}

\item{col}{Colors of CI}

\item{xlab}{Label for the x axis}

\item{ylab}{Label for the y axis}

\item{type}{Type of CI (using first order, second order or
both MSE estimators)}

\item{order_estimates}{Order of intervals in the plot}

\item{...}{Additional parameters}
}
\value{
\itemize{
\item Plot with confidence intervals for mixed effects.
}
}
\description{
Plots naive confidence intervals (CI) for mixed effects
}
\details{
The output changes slightly depending on the value
of the parameter \code{type}. If \code{type = "regular"},
naive second-order correct CIs are not present.
If \code{type = "corrected"}, naive first-order correct CIS are not present.
If \code{type = "both"}, naive first- and second-order correct CIs are present.
}
