% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.postOBSP_CI.R
\name{plot.postOBSP_CI}
\alias{plot.postOBSP_CI}
\title{Plots post-OBSP CI for mixed effects}
\usage{
\method{plot}{postOBSP_CI}(
  x,
  col = NULL,
  xlab = NULL,
  ylab = NULL,
  order_estimates = NULL,
  legend_position = NULL,
  y_axis_lim = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{postOBSP_CI}}

\item{col}{Colors of CI}

\item{xlab}{Label for the x axis}

\item{ylab}{Label for the y axis}

\item{order_estimates}{Order of intervals in the plot}

\item{legend_position}{Legend position. Default: \code{NULL}}

\item{y_axis_lim}{Limits of the y axis. Default: \code{NULL}}

\item{...}{Additional parameters}
}
\value{
\item{plot_postOBSP}{Plot with confidence intervals for mixed effects}
}
\description{
Plots post-OBSP CI for mixed effects
}
