% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postOBSP_CI.R
\name{postOBSP_CI}
\alias{postOBSP_CI}
\title{Post-OBSP confidence intervals (CI)}
\usage{
postOBSP_CI(
  X,
  y,
  clusterID,
  X_cluster_full = NULL,
  model = "NERM",
  covariate_selection_matrix = NULL,
  modelset = "all_subsets",
  intercept = FALSE,
  common = NULL,
  boot = 1000,
  alpha = 0.05
)
}
\arguments{
\item{X}{Matrix with covariates for fixed effects}

\item{y}{Vector of responses}

\item{clusterID}{Vector with cluster labels}

\item{X_cluster_full}{Matrix with cluster level covariates
for fixed effects of the full model. Default: \code{NULL}}

\item{model}{Type of mixed model: NERM, FHM, RIRS (random slopes and random intercepts)}

\item{covariate_selection_matrix}{Matrix composed of zeros and ones
indicating fixed models in each parameter. Default: \code{NULL}}

\item{modelset}{Type of model set:
\itemize{
\item \code{"all_subsets"} - all possible models using all covariates
\item \code{"nested"} - a sequence of nested models using all covariates
\item \code{"part_subset"} - all possible models using a subset of covariates
}}

\item{intercept}{Is column of ones, representing the intercept, present in X
Default: \code{TRUE}}

\item{common}{A vector indicating variables forced to be
present in each model. Default: \code{NULL}}

\item{boot}{Number of bootstrap samples.
Default: \code{boot = 1000}}

\item{alpha}{Construct 1 - alpha confidence intervals.
Default: \code{alpha = 0.05}}
}
\value{
List with elements:
\item{OBSP_min}{Index of the selected model}
\item{OBSP_models}{cAIC for all considered parameters}
\item{postOBSP_up}{Upper boundary of CI for mixed effects}
\item{postOBSP_do}{Lower boundary of CI for mixed effects}
\item{mu_hat_sel}{Mixed effects of the selected model}
}
\description{
Function \code{postOBSP_CI} provides post-OBSP confidence intervals
for mixed parameters
}
\details{
Parameter \code{boot} is needed for the calculation of the bootstrap
post-OBSP MSE of mixed effects.
}
\examples{
# Define basic parameters -------------------------------------------------
n = 15
m_i = 5
m_total = n * m_i

beta = c(2.25, -1.1, 2.43, rep(0, 2))
sig_e = 1
sig_u = 1

X = simulations_n15_mi5
X_intercept = cbind(rep(1, m_total), X)

clusterID = rep(1:n, each = m_i)

# Create responses, errors and random effects  -------------------
e_ij = rnorm(m_total, 0, sig_e)

u_i = rnorm(n, 0, sig_u)
u_i_aug = rep(u_i, each = m_i)

y = X_intercept\%*\% beta + u_i_aug + e_ij

# Post-OBSP inference ----------------------------------------

postOBSP_CI_results = postOBSP_CI(X, y,
                                  clusterID,
                                  X_cluster_full = NULL,
                                  model = "NERM",
                                  covariate_selection_matrix = NULL,
                                  modelset  = "part_subset",
                                  intercept = FALSE,
                                  common = c(1:2),
                                  boot = 250)
                                  
plot(postOBSP_CI_results)                                   

}
