% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_with_constraints.R
\name{sample_with_constraints}
\alias{sample_with_constraints}
\title{Sample with constraints}
\usage{
sample_with_constraints(
  n_starting_points,
  p,
  n,
  n_models_to_compare,
  list_constraints,
  n_samples = 10000,
  burn.in = 1000,
  scale_mvrnorm
)
}
\arguments{
\item{n_starting_points}{Number of initial starting points for sampling from truncated distribution}

\item{p}{Number of all fixed parameters under consideration (intercept included)}

\item{n}{Number of clusters (random effects)}

\item{n_models_to_compare}{Number of models in the model set to compare with}

\item{list_constraints}{List wich describes the constraints imposed on the normal distribution}

\item{n_samples}{Number of samples from constrained distribution}

\item{burn.in}{The number of burn-in iterations. The Markov chain
is sampled n_samples + burn.in times, and the last n samples are returned.}

\item{scale_mvrnorm}{Scale parameter for multivariate normal distribution to sample}
}
\description{
Function \code{sample_with_constraints} provides samples from constrained normal
distribution. Constraints are consequences of model selection.
}
\details{
\itemize{
\item \code{n_starting_points}  - in practice we use 2 times n_starting_points, because
we sample using initial values and negative initial values obtained applying function
\code{find_starting_points}
\item \code{burn.in} -  parameter of function rtmg
}
}
