% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulations_n30_mi20}
\alias{simulations_n30_mi20}
\title{Covariates to run simulations}
\format{
A data matrix with 600 rows and 4 variables generated
from the multivariate normal distribution
\code{X = rmvn(n = 30 * 20, mu = rep(0, 4), V = Omega)}
where \code{Omega} is a variance covariance matrix with 1 on the diagonal
and 0.25 elsewhere
}
\usage{
simulations_n30_mi20
}
\description{
Fixed effect covariates with 30 cluster and 20 units in each
cluster to run simulations in Claeskens, Reluga and Sperlich (2021)
}
\keyword{datasets}
