% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_information_criteria.R
\name{compute_information_criteria}
\alias{compute_information_criteria}
\title{Compute information criteria}
\usage{
compute_information_criteria(
  X,
  y,
  clusterID,
  model,
  sig_u,
  sig_e,
  fit_model_fixed,
  fit_model_mixed
)
}
\arguments{
\item{X}{Matrix with covariates for fixed effects}

\item{y}{Vector of responses}

\item{clusterID}{Vector with cluster labels}

\item{model}{Type of mixed model: NERM, FHM, RIRS (random slopes and random intercepts)}

\item{sig_u}{Variance parameter of random effects}

\item{sig_e}{Variance parameter of errors}

\item{fit_model_fixed}{Estimated model using fixed effects}

\item{fit_model_mixed}{Estimated model using fixed and random effects}
}
\value{
List with information criteria:
\item{mAIC}{Marginal AIC}
\item{mBIC}{Marginal BIC}
\item{cAIC}{Conditional AIC}
\item{deg_cAIC}{Penalty of conditional AIC}
}
\description{
Function \code{compute_infromation_criteria} provides
mAIC, cAIC and mBIC for NERM
}
\details{
Penalty term in cAIC depends on the model selected.
Function \code{compute_information_criteria} is simplified
because for now only NERM is supported.
}
\examples{

n = 10
m_i = 5
m_total = 50

clusterID = rep(1:n, m_i)
p = 10
beta = rep(2, p)
u_i = rnorm(n, 0, 2)
u_i_aug = rep(u_i, each = m_i)
X = matrix(rnorm(m_total * p), m_total, p)
y = X\%*\%beta + u_i_aug + rnorm(m_total, 0, 1)
fit <- estimate_NERM(X, y, clusterID)

IC <- compute_information_criteria(X = X, y = y,
                                   clusterID = clusterID,
                                   model = "NERM",
                                   sig_u = fit$sig_u,
                                   sig_e = fit$sig_e,
                                   fit_model_fixed = fit$fit_model_fixed,
                                   fit_model_mixed = fit$fit_model_mixed)

}
