% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_upsilon.R
\name{compute_upsilon}
\alias{compute_upsilon}
\title{Compute matrix upsilon}
\usage{
compute_upsilon(modelset_matrix)
}
\arguments{
\item{modelset_matrix}{Matrix composed of zeros and ones.}
}
\description{
Function \code{compute_upsilon} computes extended selection matrix upsilon
}
\examples{
modelset_matrix  = create_modelset("all_subsets", p = 5)
upsilon_example = compute_upsilon(modelset_matrix)

}
