% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_invK.R
\name{estimate_invK}
\alias{estimate_invK}
\title{Estimate inverse of matrix K}
\usage{
estimate_invK(X, Z, G, R)
}
\arguments{
\item{X}{Matrix with covariates for fixed effects}

\item{Z}{Matrix with covariates for random effects covariates}

\item{G}{Covariance matrix of random effects}

\item{R}{Covariance matrix of errrors}
}
\value{
Inverse of Matrix K
}
\description{
Estimate inverse of matrix K
}
