% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_NERM.R
\name{estimate_NERM}
\alias{estimate_NERM}
\title{Estimate NERM}
\usage{
estimate_NERM(X = NULL, y = NULL, clusterID = NULL, X_cluster = NULL)
}
\arguments{
\item{X}{Matrix with fixed effects covariates}

\item{y}{Vector of responses}

\item{clusterID}{Vector with cluster labels}

\item{X_cluster}{Matrix with cluster level covariates for fixed effects}
}
\value{
List with parameters
\itemize{
\item \code{fit_model_fixed} - estimated model using fixed effects
\item \code{fit_model_mixed} - estimated model using fixed and random effects
\item \code{sig_u} - variance parameter of random effects
\item \code{sig_e} - variance parameter of errors
\item \code{R} - covariance matrix of errors
\item \code{invR} - inverse of covariance matrix of errors
\item \code{detR} - determinant of covariance matrix of errors
\item \code{G} - covariance matrix of random effects
\item \code{V} - covariance matrix of response
\item \code{invV} - inverse of covariance matrix of response
\item \code{detV} - determinant of covariance matrix of response
\item \code{beta} - fixed effects (regression parameters)
\item \code{u} - random effects
\item \code{mu} - mixed effects
}
}
\description{
Estimate NERM, store fixed and random parameters and variances
}
\examples{
n = 10
m_i = 5
m_total = 50

clusterID = rep(1:5, n)
p = 10
beta = rep(2, p)
u_i = rnorm(n, 0, 2)
u_i_aug = rep(u_i, each = m_i)
X = matrix(rnorm(m_total * p), m_total, p)
y = X\%*\%beta + u_i_aug + rnorm(m_total, 0, 1)
fit_nerm <- estimate_NERM(X = X, y = y, clusterID = clusterID, X_cluster = NULL)

}
