% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results.R
\name{format_results}
\alias{format_results}
\title{Format results to plot them together}
\usage{
format_results(x, y, z, type = c("regular", "corrected", "both"))
}
\arguments{
\item{x}{An object of class \code{naive_CI}}

\item{y}{An object of class \code{postcAIC_CI}}

\item{z}{An object of class \code{postOBSP_CI}}

\item{type}{Type of naive CI (using first order, second order or
both MSE estimators)}
}
\value{
List with elements:
\itemize{
\item \code{mixed_postcAIC_CI_do} -
\item \code{mixed_postcAIC_CI_up} -
\item \code{postOBSP_do} -
\item \code{postOBSP_up} -
\item \code{mixed_naive_CI_corrected_up} - upper boundary of naive CI for mixed effects with the second order MSE
\item \code{mixed_naive_CI_corrected_do} - lower boundary of naive CI for mixed effects with the second order MSE
\item \code{mixed_naive_CI_up} - upper boundary of naive CI for mixed effects with the first order MSE
\item \code{mixed_naive_CI_do} - lower boundary of naive CI for mixed effects with the first order MSE
}
}
\description{
Function \code{format_results} creates an object of class \code{multi_CI}
to plot different confidence intervals for mixed effects together
}
\details{
The output changes slightly depending on the value
of the parameter \code{type}. If \code{type = "regular"},
naive second-order correct CIs are not present.
If \code{type = "corrected"}, naive first-order correct CIS are not present.
If \code{type = "both"}, naive first- and second-order correct CIs are present.
}
