% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_CI.R
\name{naive_CI}
\alias{naive_CI}
\title{Naive CI}
\usage{
naive_CI(
  beta_sel,
  mu_sel,
  G_sel,
  R_sel,
  V_full,
  sig_u_sel,
  sig_e_sel,
  X_full,
  Z,
  C_cluster_sel,
  clusterID,
  indices_sel = NULL,
  type_MSE_mixed = c("regular", "corrected", "both"),
  x_beta_lin_com = NULL
)
}
\arguments{
\item{beta_sel}{Fixed effects (regression parameters) of the selected model}

\item{mu_sel}{Mixed effects of the selected model}

\item{G_sel}{Covariance matrix of random effects from the selected model}

\item{R_sel}{Covariance matrix of errors from the selected model}

\item{V_full}{Covariance matrix of y from the full model}

\item{sig_u_sel}{Variance parameter of random effects of the selected model}

\item{sig_e_sel}{Variance parameter of errors of the selected model}

\item{X_full}{Matrix with a full set of covariates}

\item{Z}{Matrix with covariates for random effects}

\item{C_cluster_sel}{Matrix with cluster level covariates for fixed and random effects}

\item{clusterID}{Vector with cluster labels}

\item{indices_sel}{Indices of the selected covaraites among full covariate set}

\item{type_MSE_mixed}{Type of CI (using first order, second order or both MSE estimators)}

\item{x_beta_lin_com}{Vector or matrix to create linear combinations with
fixed parameters. Default: 'x_beta_lin_com = NULL'}
}
\value{
List with parameters
\itemize{
\item \code{beta_naive_CI_up}  - upper boundary of naive CI for fixed effects
\item \code{beta_naive_CI_do} - lower boundary of naive CI for fixed effects
\item \code{mixed_naive_CI_corrected_up} - upper boundary of naive CI for mixed effects with the second order MSE
\item \code{mixed_naive_CI_corrected_do} - lower boundary of naive CI for mixed effects with the second order MSE
\item \code{mixed_naive_CI_up} - upper boundary of naive CI for mixed effects with the first order MSE
\item \code{mixed_naive_CI_do} - lower boundary of naive CI for mixed effects with the first order MSE
\item \code{beta_x_naive_CI_up} - upper boundary of naive CI for linear combinations of fixed effects
\item \code{beta_x_naive_CI_do} - lower boundary of naive CI for linear combinations of fixed effects
}
}
\description{
Function \code{naive_CI} provides naive CI for fixed and mixed effects after model selection
}
