% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multi_CI.R
\name{plot.multi_CI}
\alias{plot.multi_CI}
\title{Plots naive, post-cAIC and post-OBSP CI for mixed effects}
\usage{
\method{plot}{multi_CI}(
  x,
  col = NULL,
  xlab = NULL,
  ylab = NULL,
  type = c("regular", "corrected", "both"),
  order_estimates = NULL,
  legend_position = NULL,
  y_axis_lim = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{multi_CI}}

\item{col}{Colors of CI. Default: \code{NULL}.}

\item{xlab}{Label for the x axis. Default: \code{NULL}.}

\item{ylab}{Label for the y axis. Default: \code{NULL}.}

\item{type}{Type of CI (using first order, second order or
both MSE estimators)}

\item{order_estimates}{Order of intervals in the plot. Default: \code{NULL}.}

\item{legend_position}{Legend position. Default: \code{NULL}}

\item{y_axis_lim}{Limits of the y axis. Default: \code{NULL}}

\item{...}{Additional parameters}
}
\value{
\item{plot_multi}{Plot with confidence intervals for mixed effects}
}
\description{
Plots naive, post-cAIC and post-OBSP CI for mixed effects
}
