% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.postcAIC_CI.R
\name{plot.postcAIC_CI}
\alias{plot.postcAIC_CI}
\title{Plots post-cAIC CI for mixed effects}
\usage{
\method{plot}{postcAIC_CI}(x, col = NULL, xlab = NULL, ylab = NULL, order_estimates = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{postcAIC_CI}}

\item{col}{Colour of CI}

\item{xlab}{Label for the x axis}

\item{ylab}{Label for the y axis}

\item{order_estimates}{Order of intervals in the plot}

\item{...}{Additional parameters}
}
\description{
Plots post-cAIC CI for mixed effects
}
