% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcAIC_CI.R
\name{postcAIC_CI}
\alias{postcAIC_CI}
\title{Post-cAIC confidence intervals}
\usage{
postcAIC_CI(
  cAIC_min,
  degcAIC_models,
  X_full,
  X_cluster_full,
  sig_u_full,
  sig_e_full,
  model = "NERM",
  clusterID,
  beta_sel,
  mu_sel,
  modelset_matrix,
  x_beta_lin_com = NULL,
  n_starting_points = 5,
  scale_mvrnorm = 1,
  alpha = 0.05
)
}
\arguments{
\item{cAIC_min}{Index of the selected model among models in the modelset}

\item{degcAIC_models}{Penalty for all considered models}

\item{X_full}{Matrix with a full set of covariates}

\item{X_cluster_full}{Matrix with cluster level covariates for fixed effects of the full model}

\item{sig_u_full}{Variance parameter of random effects from the full model}

\item{sig_e_full}{Variance parameter of errors from the full model}

\item{model}{Type of mixed model: NERM, FHM, RIRS (random slopes and random intercepts).}

\item{clusterID}{Vector with cluster labels}

\item{beta_sel}{Fixed effects (regression parameters) of the selected model}

\item{mu_sel}{Mixed effects of the selected model}

\item{modelset_matrix}{Matrix composed of zeros and ones.}

\item{x_beta_lin_com}{Vector or matrix to create linear combinations with
fixed parameters. Default: 'x_beta_lin_com = NULL'}

\item{n_starting_points}{Number of initial starting points for sampling from truncated distribution}

\item{scale_mvrnorm}{Scale parameter for multivariate normal distribution to sample}

\item{alpha}{Construct 1 - alpha confidence intervals}
}
\value{
List with parameters
\itemize{
\item \code{beta_PoSI_CI_up}  - upper boundary of CI for fixed effects
\item \code{beta_PoSI_CI_do} - lower boundary of CI for fixed effects
\item \code{mixed_PoSI_CI_up} - upper boundary of CI for mixed effects
\item \code{mixed_PoSI_CI_do} - lower boundary of CI for mixed effects
\item \code{beta_x_PoSI_CI_up} - upper boundary of CI for linear combinations of fixed effects
\item \code{beta_x_PoSI_CI_do} - lower boundary of CI for linear combinations of fixed effects
}
}
\description{
Function \code{postcAIC_CI} provides post-cAIC confidence intervals for
mixed and fixed effects under NERM
}
\examples{
n = 10
m_i = 5
m_total = 50

clusterID = rep(1:n, m_i)
p = 10
beta = rep(2, p)
u_i = rnorm(n, 0, 2)
u_i_aug = rep(u_i, each = m_i)
X = matrix(rnorm(m_total * p), m_total, p)
y = X\%*\%beta + u_i_aug + rnorm(m_total, 0, 1)

cAIC_model_set =
compute_cAIC_for_model_set(X, y, clusterID,
                           model = "NERM",
                           covariate_selection_matrix = NULL,
                           modelset  = "part_subset",
                           common = c(1:8),
                           intercept = FALSE)

cAIC_min = cAIC_model_set$cAIC_min
degcAIC_models = cAIC_model_set$degcAIC_models
X_full = cAIC_model_set$X_full
X_cluster_full = cAIC_model_set$X_cluster_full

sig_u_full = cAIC_model_set$sig_u_full
sig_e_full = cAIC_model_set$sig_u_full

beta_sel = cAIC_model_set$beta_sel
mu_sel = cAIC_model_set$mu_sel

modelset_matrix = cAIC_model_set$modelset_matrix
x_beta_lin_com = cAIC_model_set$X_cluster_full

postcAIC_CI_results = postcAIC_CI(cAIC_min,
                                 degcAIC_models,

                                 X_full,
                                 X_cluster_full,
                                 sig_u_full,
                                 sig_e_full,
                                 model = "NERM",
                                 clusterID,

                                 beta_sel,
                                 mu_sel,

                                 modelset_matrix,
                                 x_beta_lin_com = NULL)


}
