% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcAIC_CI.R
\name{postcAIC_CI}
\alias{postcAIC_CI}
\title{Post-cAIC confidence intervals}
\usage{
postcAIC_CI(
  cAIC_min,
  degcAIC_models,
  Z,
  X_full,
  X_cluster_full,
  G_full,
  R_full,
  V_full,
  beta_sel,
  mu_sel,
  modelset = "all_subsets",
  modelset_matrix,
  x_beta_lin_com = NULL,
  n_starting_points = 5,
  scale_mvrnorm
)
}
\arguments{
\item{cAIC_min}{Index of the selected model among models in the modelset}

\item{degcAIC_models}{Penalty for all considered models}

\item{Z}{Matrix with covariates for random effects covariates}

\item{X_full}{Matrix with a full set of covariates}

\item{X_cluster_full}{Matrix with cluster level covariates for fixed effects of the full model}

\item{G_full}{Covariance matrix of random effects of the full model}

\item{R_full}{Covariance matrix of errors of the full model}

\item{V_full}{Covariance matrix of response of the full model}

\item{beta_sel}{Fixed effects (regression parameters) of the selected model}

\item{mu_sel}{Mixed effects of the selected model}

\item{modelset}{Type of model set:
\itemize{
\item \code{"all_subsets"} - all possible models for all parameters
\item \code{"nested"} - a sequence of nested models for all parameters
\item \code{"part_subset"} - all possible models for a subset of parameters
}}

\item{modelset_matrix}{Matrix composed of zeros and ones.}

\item{x_beta_lin_com}{Vector or matrix to create linear combinations with
fixed parameters. Default: 'x_beta_lin_com = NULL'}

\item{n_starting_points}{Number of initial starting points for sampling from truncated distribution}

\item{scale_mvrnorm}{Scale parameter for multivariate normal distribution to sample}
}
\value{
List with parameters
\itemize{
\item \code{cAIC_min}  - Index of selected model
\item \code{degcAIC_models} - Penalty for all considered models
\item \code{beta_sel} - fixed effects (regression parameters) of the selected model
\item \code{mu_sel} - mixed effects of the selected model
\item \code{V_sel} - covariance matrix of response of the selected model
\item \code{G_sel} - covariance matrix of random effects of the selected model
\item \code{R_sel} - covariance matrix of response of the selected model
\item \code{X_sel} - matrix with fixed effects covariates of the selected model
\item \code{indices_sel} - indices of the selected covaraites among full covariate set
\item \code{sig_u_sel} - variance parameter of random effects of the selected model
\item \code{sig_e_sel} - variance parameter of errors of the selected model
\item \code{V_sel} -  covariance matrix of response of the selected model
\item \code{invV_full} - inverse of covariance matrix of response of the full model
\item \code{G_full} - covariance matrix of random effects of the selected model
\item \code{R_full} - covariance matrix of response of the full model
\item \code{X_full} - matrix with fixed effects covariates of the full model
\item \code{X_cluster_full} - matrix with cluster level covariates for fixed effects of the full model
\item \code{Z} - matrix of covariates for random effects
\item \code{modelset_matrix} -  matrix composed of zeros and ones. Ones correspond to
parameters in a model which is represented in nth row.
}

List with parameters
\itemize{
\item \code{beta_PoSI_CI_up}  - upper boundary of CI for fixed effects
\item \code{beta_PoSI_CI_do} - lower boundary of CI for fixed effects
\item \code{mixed_PoSI_CI_up} - upper boundary of CI for mixed effects
\item \code{mixed_PoSI_CI_do} - lower boundary of CI for mixed effects
\item \code{beta_x_PoSI_CI_up} - upper boundary of CI for linear combinations of fixed effects
\item \code{beta_x_PoSI_CI_do} - lower boundary of CI for linear combinations of fixed effects
}
}
\description{
Function \code{postcAIC_CI} provides post-cAIC confidence intervals for
mixed and fixed effects under NERM
}
