% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{postcAIC_nhaens}
\alias{postcAIC_nhaens}
\title{Subset of data from R package NHAENS}
\format{
A data frame with 3009 rows and 9 columns:
\describe{
\item{\code{PhysActive}}{Does participant do sport, fitness or recreational activities? (Yes or No)}
\item{\code{CurrentSmokingStatus}}{Derived variable indicating if person is smoking (1) or is not smoking (0)}
\item{\code{Diabetes}}{Is person suffering from diabetes? (Yes or No)}
\item{\code{SleepHrsNight}}{Number of sleeping hours (numeric value)}
\item{\code{BPSys2}}{Systolic blood pressure (numeric value)}
\item{\code{DirectChol}}{Level of direct cholesterol (numeric value)}
\item{\code{Poverty}}{Numeric indicator with smaller values indicating more poverty}
\item{\code{clusterID}}{Cluster indicator. Cross-section of variables age, biological sex and entinicity}
\item{\code{log_BMI}}{Natural logarithm of Body Mass Index (BMI)}
}
}
\source{
\url{https://cran.r-project.org/web/packages/NHANES/}
\url{https://github.com/KatarzynaReluga/postcAIC}.
}
\usage{
postcAIC_nhaens
}
\description{
Data set used in Claeskens, Reluga and Sperlich (2021).
}
\keyword{datasets}
