% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postOBSP_CI.R
\name{postOBSP_CI}
\alias{postOBSP_CI}
\title{Post OBSP confidence intervals}
\usage{
postOBSP_CI(
  X_full,
  y,
  clusterID,
  Z,
  X_cluster_full,
  G_full,
  V_full,
  R_full,
  sig_u_full,
  sig_e_full,
  modelset_matrix,
  boot = 1000
)
}
\arguments{
\item{X_full}{Matrix with a full set of covariates}

\item{y}{Vector of responses}

\item{clusterID}{Vector with cluster labels}

\item{Z}{Matrix with covariates for random effects covariates}

\item{X_cluster_full}{Matrix with cluster level covariates for fixed effects of the full model}

\item{G_full}{Covariance matrix of random effects of the full model}

\item{V_full}{Covariance matrix of response of the full model}

\item{R_full}{Covariance matrix of errors of the full model}

\item{sig_u_full}{Variance parameter of random effects from the full model}

\item{sig_e_full}{Variance parameter of errors from the full model}

\item{modelset_matrix}{Matrix composed of zeros and ones.}

\item{boot}{Number of boostrap samples}
}
\description{
Function \code{postOBSP_CI} provides post-OBSP confidence intervals for mixed parameters
}
\details{
\itemize{
\item The parameter \code{boot} is needed for the calculation of the boostrap post-OBSP MSE of mixed effects
}
}
