% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize_Sigma.R
\name{vectorize_Sigma}
\alias{vectorize_Sigma}
\title{Vectorize Sigma}
\usage{
vectorize_Sigma(Sigma, index_cov_terms, p, index_full_model)
}
\arguments{
\item{Sigma}{Matrix Sigma}

\item{index_cov_terms}{Index of covariance terms (for example,
12, 34, etc)}

\item{p}{Number of all fixed parameters under consideration (intercept included)}

\item{index_full_model}{Index of a full model in the model set}
}
\description{
Function \code{vectorise_Sigma} vectorizes and orders
unique elements of variance matrix Sigma
}
\keyword{internal}
